Splits GroupsRedux and PhotoViewingRedux; Makes GroupsRedux cachable in the persisted redux
