+++ App/Redux/RootReducer.ts
@@ -10,8 +10,9 @@ import {
 import { accountReducer } from '../features/account'
 import { reducer as authReducer } from './AuthRedux'
 import { contactsReducer } from '../features/contacts'
-import { reducer as photoViewingReducer } from './PhotoViewingRedux'
+import { reducer as groupsReducer } from './GroupsRedux'
 import { reducer as prefrencesReducer } from './PreferencesRedux'
+import { reducer as photoViewingReducer } from './PhotoViewingRedux'
 import { reducer as notificationsReducer } from './NotificationsRedux'
 import { reducer as threadsReducer } from './ThreadsRedux'
 import { reducer as uiReducer } from './UIRedux'
@@ -370,7 +371,7 @@ const persistConfig: PersistConfig = {
   key: 'primary',
   storage: AsyncStorage,
   version: 23,
-  whitelist: ['account', 'preferences', 'deviceLogs'],
+  whitelist: ['account', 'preferences', 'deviceLogs', 'groups'],
   migrate: createMigrate(migrations, { debug: false }),
   debug: false
 }
@@ -378,8 +379,9 @@ const persistConfig: PersistConfig = {
 const rootReducer = combineReducers({
   auth: authReducer,
   contacts: contactsReducer,
-  photoViewing: photoViewingReducer,
+  groups: groupsReducer,
   preferences: prefrencesReducer,
+  photoViewing: photoViewingReducer,
   notifications: notificationsReducer,
   threads: threadsReducer,
   ui: uiReducer,
