+++ App/SB/components/ThreadSelect/index.tsx
@@ -7,8 +7,8 @@ import { View, Text, FlatList, ListRenderItemInfo } from 'react-native'
 import { RootState, RootAction } from '../../../Redux/Types'
 import ThreadSelectCard from './ThreadSelectCard'
 import ThreadCreateCard from './ThreadCreateCard'
-import { getSharedThreads } from '../../../Redux/PhotoViewingSelectors'
-import { ThreadData } from '../../../Redux/PhotoViewingRedux'
+import { getSharedThreads } from '../../../Redux/GroupsSelectors'
+import { ThreadData } from '../../../Redux/GroupsRedux'
 import UIActions from '../../../Redux/UIRedux'
 
 import styles from './statics/styles'
