+++ App/Sagas/PhotoViewingSagas.ts
@@ -4,17 +4,20 @@ import { ActionType, getType } from 'typesafe-actions'
 import uuid from 'uuid/v4'
 import Config from 'react-native-config'
 
+import GroupsActions from '../Redux/GroupsRedux'
 import PhotoViewingActions from '../Redux/PhotoViewingRedux'
 import ThreadsActions, { InboundInvite } from '../Redux/ThreadsRedux'
 import { inboundInviteByThreadName } from '../Redux/ThreadsSelectors'
 import TextileEventsActions from '../Redux/TextileEventsRedux'
 import UIActions from '../Redux/UIRedux'
 import {
-  photoAndComment,
+  photoAndComment
+} from '../Redux/PhotoViewingSelectors'
+import {
   shouldNavigateToNewThread,
   shouldSelectNewThread,
   photoToShareToNewThread
-} from '../Redux/PhotoViewingSelectors'
+} from '../Redux/GroupsSelectors'
 import Textile, {
   IThread,
   IAddThreadConfig,
@@ -29,8 +32,8 @@ import { shareWalletImage } from './ImageSharingSagas'
 export function* monitorNewThreadActions() {
   while (true) {
     const action: ActionType<
-      typeof PhotoViewingActions.threadAdded
-    > = yield take(getType(PhotoViewingActions.threadAdded))
+      typeof GroupsActions.threadAdded
+    > = yield take(getType(GroupsActions.threadAdded))
     const { id, name } = action.payload
     const photoToShare:
       | { threadName: string; imageId: string; comment?: string }
@@ -42,7 +45,7 @@ export function* monitorNewThreadActions() {
       name
     )
 
-    yield put(PhotoViewingActions.clearNewThreadActions())
+    yield put(GroupsActions.clearNewThreadActions())
 
     if (photoToShare && photoToShare.threadName === name) {
       const { imageId, comment } = photoToShare
@@ -63,14 +66,14 @@ export function* monitorNewThreadActions() {
 }
 
 export function* monitorThreadAddedNotifications(
-  action: ActionType<typeof PhotoViewingActions.threadAddedNotification>
+  action: ActionType<typeof GroupsActions.threadAddedNotification>
 ) {
   try {
     // We need this one because the callback we get from the node doesn't include key. This queries for the thread and gets
     // all the required data for threadAdded()
     const thread: IThread = yield call(Textile.threads.get, action.payload.id)
     const withValid = { ...thread, valid: true }
-    yield put(PhotoViewingActions.threadAdded(withValid))
+    yield put(GroupsActions.threadAdded(withValid))
   } catch (error) {
     yield put(
       TextileEventsActions.newErrorMessage(
@@ -78,12 +81,12 @@ export function* monitorThreadAddedNotifications(
         error.message
       )
     )
-    yield put(PhotoViewingActions.addThreadError(error))
+    yield put(GroupsActions.addThreadError(error))
   }
 }
 
 export function* addThread(
-  action: ActionType<typeof PhotoViewingActions.addThreadRequest>
+  action: ActionType<typeof GroupsActions.addThreadRequest>
 ) {
   const { name, whitelist, type, sharing } = action.payload
   const { invites } = action.meta
@@ -109,12 +112,12 @@ export function* addThread(
     }
   } catch (error) {
     yield put(TextileEventsActions.newErrorMessage('addThread', error.message))
-    yield put(PhotoViewingActions.addThreadError(error))
+    yield put(GroupsActions.addThreadError(error))
   }
 }
 
 export function* removeThread(
-  action: ActionType<typeof PhotoViewingActions.removeThreadRequest>
+  action: ActionType<typeof GroupsActions.removeThreadRequest>
 ) {
   const { id } = action.payload
   try {
@@ -124,12 +127,12 @@ export function* removeThread(
     yield put(
       TextileEventsActions.newErrorMessage('removeThread', error.message)
     )
-    yield put(PhotoViewingActions.removeThreadError(error))
+    yield put(GroupsActions.removeThreadError(error))
   }
 }
 
 export function* refreshThreads(
-  action: ActionType<typeof PhotoViewingActions.refreshThreadsRequest>
+  action: ActionType<typeof GroupsActions.refreshThreadsRequest>
 ) {
   try {
     const threadsResult: IThreadList = yield call(Textile.threads.list)
@@ -143,20 +146,20 @@ export function* refreshThreads(
       if (useIt) {
         const valid = thread.headBlocks.length > 0
         const withValid = { ...thread, valid }
-        yield put(PhotoViewingActions.insertThread(withValid))
-        yield put(PhotoViewingActions.refreshThreadRequest(thread.id))
+        yield put(GroupsActions.insertThread(withValid))
+        yield put(GroupsActions.refreshThreadRequest(thread.id))
       }
     }
   } catch (error) {
     yield put(
       TextileEventsActions.newErrorMessage('refreshThreads', error.message)
     )
-    yield put(PhotoViewingActions.refreshThreadsError(error))
+    yield put(GroupsActions.refreshThreadsError(error))
   }
 }
 
 export function* refreshThread(
-  action: ActionType<typeof PhotoViewingActions.refreshThreadRequest>
+  action: ActionType<typeof GroupsActions.refreshThreadRequest>
 ) {
   const { threadId } = action.payload
   try {
@@ -167,10 +170,10 @@ export function* refreshThread(
       -1
     )
     yield put(
-      PhotoViewingActions.refreshThreadSuccess(threadId, photosResult.items)
+      GroupsActions.refreshThreadSuccess(threadId, photosResult.items)
     )
   } catch (error) {
-    yield put(PhotoViewingActions.refreshThreadError(threadId, error))
+    yield put(GroupsActions.refreshThreadError(threadId, error))
   }
 }
