+++ App/features/group/rename-group/sagas.ts
@@ -2,7 +2,7 @@ import { ActionType, getType } from 'typesafe-actions'
 import { call, put, takeEvery, all } from 'redux-saga/effects'
 import Textile from '@textile/react-native-sdk'
 import { renameGroup } from './actions'
-import PhotoViewingActions from '../../../Redux/PhotoViewingRedux'
+import GroupsActions from '../../../Redux/GroupsRedux'
 
 export function* handleRenameGroupRequest(
   action: ActionType<typeof renameGroup.request>
@@ -10,7 +10,7 @@ export function* handleRenameGroupRequest(
   const { threadId, name } = action.payload
   try {
     yield call(Textile.threads.rename, threadId, name)
-    yield put(PhotoViewingActions.updateThreadName(threadId, name))
+    yield put(GroupsActions.updateThreadName(threadId, name))
     yield put(renameGroup.success({ threadId }))
   } catch (error) {
     yield put(renameGroup.failure({ threadId, error }))
